fixed Transaction

+++ src/main/java/com/enterpriseintellijence/enterpriseintellijence/data/services/TransactionServiceImp.java
@@ -1,14 +1,26 @@
 package com.enterpriseintellijence.enterpriseintellijence.data.services;
 
+import com.enterpriseintellijence.enterpriseintellijence.data.entities.Order;
+import com.enterpriseintellijence.enterpriseintellijence.data.entities.PaymentMethod;
 import com.enterpriseintellijence.enterpriseintellijence.data.entities.Transaction;
+import com.enterpriseintellijence.enterpriseintellijence.data.entities.User;
+import com.enterpriseintellijence.enterpriseintellijence.data.repository.OrderRepository;
+import com.enterpriseintellijence.enterpriseintellijence.data.repository.PaymentMethodRepository;
 import com.enterpriseintellijence.enterpriseintellijence.data.repository.TransactionRepository;
 import com.enterpriseintellijence.enterpriseintellijence.dto.TransactionDTO;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.Availability;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.OrderState;
+import com.enterpriseintellijence.enterpriseintellijence.dto.enums.UserRole;
 import com.enterpriseintellijence.enterpriseintellijence.exception.IdMismatchException;
+import com.enterpriseintellijence.enterpriseintellijence.security.JwtContextUtils;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.time.Clock;
+import java.time.LocalDateTime;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -16,37 +28,67 @@ import java.util.stream.Collectors;
 public class TransactionServiceImp implements TransactionService{
 
     private final TransactionRepository transactionRepository;
+    private final OrderRepository orderRepository;
+    private final PaymentMethodRepository paymentMethodRepository;
     private final ModelMapper modelMapper;
-    public TransactionDTO createTransaction(TransactionDTO transactionDTO) {
+    private final JwtContextUtils jwtContextUtils;
+    private final Clock clock;
+
+    public TransactionDTO createTransaction(TransactionDTO transactionDTO) throws IllegalAccessException {
         Transaction transaction = mapToEntity(transactionDTO);
+        User loggedUser = jwtContextUtils.getUserLoggedFromContext();
+
+        if(loggedUser!=null && !transaction.getOrder().getUser().getId().equals(loggedUser.getId()))
+            throw new IllegalAccessException("Cannot create transaction");
+        checkCardOwnership(loggedUser,transaction.getPaymentMethod());
+
+        Order order = orderRepository.getReferenceById(transaction.getOrder().getId());
+        order.setState(OrderState.PURCHASED);
+        order.getProduct().setAvailability(Availability.UNAVAILABLE);
+        transaction.setOrder(order);
+        transaction.setCreationTime(LocalDateTime.now(clock));
         transaction = transactionRepository.save(transaction);
         return mapToDTO(transaction);
     }
 
-    public TransactionDTO replaceTransaction(String id, TransactionDTO transactionDTO) {
-        throwOnIdMismatch(id, transactionDTO);
-        Transaction oldTransaction = transactionRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-
-        // todo: implementare
-
-
-        Transaction newTransaction = transactionRepository.save(mapToEntity(transactionDTO));
-        return mapToDTO(newTransaction);
-        //todo: da fare perchè non c'è l'allacciamento tra Transaction e User
+    public TransactionDTO replaceTransaction(String id, TransactionDTO transactionDTO) throws IllegalAccessException {
+        return updateTransaction(id,transactionDTO);
     }
 
-    public TransactionDTO updateTransaction(String id, TransactionDTO patch)  {
-        TransactionDTO transaction = mapToDTO(transactionRepository.findById(id).orElseThrow(EntityNotFoundException::new));
-        transaction.setAmount(patch.getAmount());
-        transaction.setPaymentMethod(patch.getPaymentMethod());
-        transactionRepository.save(mapToEntity(transaction));
-        return transaction;
-    }
-
-    public void deleteTransaction(String id) {
+    public TransactionDTO updateTransaction(String id, TransactionDTO patch) throws IllegalAccessException {
         Transaction transaction = transactionRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+
+        //check sugli id
+        throwOnIdMismatch(id,patch);
+
+        User loggedUser = jwtContextUtils.getUserLoggedFromContext();
+        //controllo che chi effettua l'operazione sia un admin o il proprietario della transazione
+        if(loggedUser!= null && loggedUser.getRole().equals(UserRole.USER) && !transaction.getOrder().getUser().getId().equals(loggedUser.getId()))
+            throw new IllegalAccessException("Cannot modify transaction");
+        //controllo che il campo del metodo non sia null
+        if(patch.getPaymentMethod()!=null){
+            PaymentMethod paymentMethod = paymentMethodRepository.getReferenceById(patch.getPaymentMethod().getId());
+            //controllo che il proprietario della carta sia nei metodi di pagamento dell'utente
+            checkCardOwnership(transaction.getOrder().getUser(),paymentMethod);
+            //verifico se il metodo di pagamento è effettivamente diverso
+            if(!transaction.getPaymentMethod().equals(paymentMethod)) {
+                transaction.setPaymentMethod(paymentMethod);
+                transaction.getOrder().setState(OrderState.PURCHASED);
+                transaction.getOrder().getProduct().setAvailability(Availability.UNAVAILABLE);
+                transactionRepository.save(transaction);
+            }
+        }
+
+        return mapToDTO(transaction);
+    }
+
+    public void deleteTransaction(String id) throws IllegalAccessException {
+        Transaction transaction = transactionRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        User loggedUser = jwtContextUtils.getUserLoggedFromContext();
+        //controllo che chi effettua l'operazione sia un admin o il proprietario della transazione
+        if(loggedUser.getRole().equals(UserRole.USER) && !transaction.getOrder().getUser().getId().equals(loggedUser.getId()))
+            throw new IllegalAccessException("Cannot modify transaction");
         transactionRepository.deleteById(id);
-        mapToDTO(transaction);
     }
 
     public TransactionDTO transactionById(String id) {
@@ -66,4 +108,12 @@ public class TransactionServiceImp implements TransactionService{
 
     public Transaction mapToEntity(TransactionDTO transactionDTO){return modelMapper.map(transactionDTO,Transaction.class);}
     public TransactionDTO mapToDTO(Transaction transaction){return modelMapper.map(transaction,TransactionDTO.class);}
+
+    private boolean checkCardOwnership(User user, PaymentMethod paymentMethod) throws IllegalAccessException {
+        if(!user.getPaymentMethods().contains(paymentMethod))
+            throw new IllegalAccessException("Cannot use this credit card");
+
+        return true;
+
+    }
 }
